Fixed JSON saves for guild joining
"servers" wasn't saved when the new guild data was added, breaking the entire structure.

When the bot joins a new guild, it checks to see if it already has data for it.

Removed joining message about setting up the bot as I couldn't find a way to get the default channel.