Author variable made for sent message. Cache per person redesign
Author variable made for more efficient referral to the author.

Add xp was made into a function due to the redesign of the cache where each member has their own cooldown per guild. getGuildRanks had to be made into its own function so it could be referred to in multiple places now addXP was in its own function.

On ready info changed to say "starting" instead of "finished" when starting and time to start is given.