Created debug level option
Before you say anything, I know this belongs in main and not in some random branch I was working in, but I wanted to discuss one of the changes you made to main instead of reverting it without your permission but I'm too tired to do it so instead I thought I'd add this here.
Apologies in advance.