Removed upgrade_json.py
The idea of getting an auto-updater to go through a file and replace only certain parts of it with new functional JSON code is a bit too advanced and would require its own interface. Factors such as what variables are optional for servers, or a common server data structure would need to be specified and we're not at that stage yet.
The problem of having to update the data file manually every time its structure is changed still persists, so a solution is still needed. However since this is a solution needed for the purposes of bot development and not a solution to any problems the bot is posing in and of itself, I think we can safely discard it for now.
In short, whenever you need to change the data file's structure, you'd best go through each server one by one very carefully. If that made you sigh in desperation then consider the hassle of writing a function that did this for you instead of working on the bot directly, and then be grateful that I stopped you.