Replaced command_resetter.py and clear_slash_command.py with hmth_tools.py
Helminth had (almost) every API request implemented with a logging wrapper, so I figured I'd bring that here.
Any code you want to run you should add in the "if name equals main" indentation at the bottom, and always discard changes to this file when you're finished. For future reference, there is a way of telling git to `assume-unchanged`.